+++ src/plugin.ts
@@ -8,8 +8,7 @@ import {
 } from "./actions/rs3-gim-group-rank";
 import { Rs3LevelTracker } from "./actions/rs3-level-tracker";
 import { Rs3SkillTracker } from "./actions/rs3-skill-tracker";
-import { Rs3WildyEventTimer } from "./actions/rs3-wildy-event-timer";
-import { Rs3Goldberg } from "./actions/rs3-goldberg";
+import { Rs3DailyXpGain } from "./actions/rs3-daily-xp-gain";
 import { startUpdateScheduler } from "./update-scheduler";
 import { checkForUpdate } from "./update-check";
 
@@ -20,8 +19,7 @@ streamDeck.actions.registerAction(new Rs3GimGroupRank());
 streamDeck.actions.registerAction(new Rs3GimGroupRankOther());
 streamDeck.actions.registerAction(new Rs3LevelTracker());
 streamDeck.actions.registerAction(new Rs3SkillTracker());
-streamDeck.actions.registerAction(new Rs3WildyEventTimer());
-streamDeck.actions.registerAction(new Rs3Goldberg());
+streamDeck.actions.registerAction(new Rs3DailyXpGain());
 startUpdateScheduler(checkForUpdate, (message, data) =>
 	streamDeck.logger.info(`${message} ${JSON.stringify(data ?? {})}`)
 );